+++ RELEASE_NOTES.md
@@ -2,6 +2,12 @@ See the [CHANGELOG](CHANGELOG.md) for a complete list of changes.
 
 `porter` is [semantically versioned](http://semver.org/spec/v2.0.0.html)
 
+v4.8
+====
+
+Some [HAProxy timeouts](docs/detailed_design/config-reference.md#timeout)
+are configurable.
+
 v4.7
 ====